refactor(apibuilder): refactor apiBUilder class
