+++ src/pages/App/App.tsx
@@ -1,23 +1,19 @@
 import './App.css';
-import { ProductDiscount } from '@commercetools/platform-sdk';
 import SharedLayout from 'pages/App/layouts/SharedLayout/SharedLayout';
 import PrivateRoute from 'pages/App/routes/PrivateRoute/PrivateRoute';
 import RestrictedRoute from 'pages/App/routes/RestrictedRoute/RestrictedRoute';
 import IUser from 'pages/App/types/interfaces/IUser';
 import CatalogPage from 'pages/CatalogPage/CatalogPage';
 import Profile from 'pages/ProfilePage/Profile';
-import { lazy, useEffect, useState } from 'react';
+import { lazy, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import { setCredentials } from 'shared/api/authApi/store/authSlice';
-import { ApiBuilder } from 'shared/libs/commercetools/apiBuilder';
-import setProductsArray from 'shared/utils/setProductsArray.ts';
+import { currentClient } from 'shared/libs/commercetools/apiBuilder';
 import Loader from 'widgets/Loader/Loader';
 
 import 'react-toastify/dist/ReactToastify.css';
-
-import IProductData from './types/interfaces/IProductData.ts';
 
 const NotFoundPage = lazy(() => import('pages/NotFoundPage/NotFound'));
 const SignInPage = lazy(() => import('pages/SignInPage/SignIn'));
@@ -27,11 +23,6 @@ const MainPage = lazy(() => import('pages/MainPage/Main'));
 function App() {
   const dispatch = useDispatch();
   const isRefreshing = false;
-  const currentClient = new ApiBuilder();
-  const productsList: IProductData[] = [];
-  const discountsList: ProductDiscount[] = [];
-  const [products, setProducts] = useState(productsList);
-  const [discounts, setDiscounts] = useState(discountsList);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -61,16 +52,6 @@ function App() {
       } else {
         await currentClient.createAnonymousClient();
       }
-
-      await currentClient
-        .getProducts()
-        .then((resp) => resp?.body.results)
-        .then((resp) => setProducts(setProductsArray(resp)));
-
-      await currentClient
-        .getProductsDiscount()
-        .then((resp) => resp?.body.results)
-        .then((resp) => setDiscounts(resp as ProductDiscount[]));
     };
 
     fetchData();
@@ -96,12 +77,7 @@ function App() {
                 />
               )}
             />
-            <Route
-              path="/catalog"
-              element={
-                <CatalogPage products={products} discounts={discounts} />
-              }
-            />
+            <Route path="/catalog" element={<CatalogPage />} />
 
             <Route
               path="/signin"
