+++ src/pages/CatalogPage/CatalogPage.tsx
@@ -1,16 +1,18 @@
 import { ProductDiscount } from '@commercetools/platform-sdk';
 import IProductData from 'pages/App/types/interfaces/IProductData';
+import './CatalogPage.modules.css';
+import { useEffect, useState } from 'react';
+import { currentClient } from 'shared/libs/commercetools/apiBuilder';
+import setProductsArray from 'shared/utils/setProductsArray';
 import FilterAside from 'widgets/FilterAside/FilterAside';
 import ProductCard from 'widgets/ProductCard/ProductCard';
-import './CatalogPage.modules.css';
 
-function CatalogPage({
-  products,
-  discounts,
-}: {
-  products: IProductData[];
-  discounts: ProductDiscount[];
-}) {
+function CatalogPage() {
+  const productsList: IProductData[] = [];
+  const discountsList: ProductDiscount[] = [];
+  const [products, setProducts] = useState(productsList);
+  const [discounts, setDiscounts] = useState(discountsList);
+
   function getDiscont(name: string | undefined) {
     let discont: ProductDiscount | boolean = false;
 
@@ -22,6 +24,23 @@ function CatalogPage({
 
     return discont;
   }
+  useEffect(() => {
+    const fetchData = async () => {
+      await currentClient
+        .getProducts()
+        .then((resp) => resp?.body.results)
+        .then((resp) => setProducts(setProductsArray(resp)));
+
+      await currentClient
+        .getProductsDiscount()
+        .then((resp) => resp?.body.results)
+        .then((resp) => setDiscounts(resp as ProductDiscount[]));
+    };
+
+    if (!products.length) {
+      fetchData();
+    }
+  }, [products]);
 
   return (
     <div className="catalog-page">
