+++ src/pages/SignUpPage/SignUp.test.tsx
@@ -1,34 +1,32 @@
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import store from 'shared/api/authApi/store/store.ts';
-import { ApiBuilder } from 'shared/libs/commercetools/apiBuilder.ts';
 import { render, screen } from 'shared/utils/test-utils';
 import { describe, expect, it } from 'vitest';
 
 import SignUp from './SignUp.tsx';
+import { currentClient } from 'shared/libs/commercetools/apiBuilder.ts';
 
 describe('Simple working test', () => {
-  const client = new ApiBuilder();
-
   it('the link is visible', () => {
     render(
       <Provider store={store}>
         <BrowserRouter basename="/">
-          <SignUp client={client} />
+          <SignUp client={currentClient} />
         </BrowserRouter>
-      </Provider>,
+      </Provider>
     );
     expect(
-      screen.getByText('Already have an account? Sign In'),
+      screen.getByText('Already have an account? Sign In')
     ).toBeInTheDocument();
   });
   it('the input is visible', () => {
     render(
       <Provider store={store}>
         <BrowserRouter basename="/">
-          <SignUp client={client} />
+          <SignUp client={currentClient} />
         </BrowserRouter>
-      </Provider>,
+      </Provider>
     );
     expect(screen.getByText('Password')).toBeInTheDocument();
   });
